Debugging process 2

+++ srcJava/com/rti/perftest/harness/PerfTest.java
@@ -244,7 +244,7 @@ public final class PerfTest {
                     sb.append(_packetsPerSecHistory.get(i) + " , ");
                     sb.append(_throughputHistory.get(i) + " , ");
                     sb.append(_lostHistory.get(i) + " , ");
-                    sb.append(((_lostHistory.get(i) / _packetsHistory.get(i))) + " , ");
+                    sb.append((double)(_lostHistory.get(i) / _packetsHistory.get(i)) + " , ");
                     sb.append('\n');
 
                     testVariable = (double)(_lostHistory.get(i) / _packetsHistory.get(i));
@@ -254,7 +254,7 @@ public final class PerfTest {
                     throughputTotal += _throughputHistory.get(i);
                     packetsReceivedTotal += _packetsHistory.get(i);
                     packetsPerSecTotal += _packetsPerSecHistory.get(i);
-                    lostPercentTotal += (_lostHistory.get(i) / _packetsHistory.get(i));
+                    lostPercentTotal += (double)(_lostHistory.get(i) / _packetsHistory.get(i));
                     throughputCount++;
 
                     System.out.println("lostPercentTotal: " + lostPercentTotal);
